requirements.txt
# Core TTS
qwen-tts

# Audio I/O
soundfile

# Tensor / ML
torch
torchvision
torchaudio

# Web UI
gradio

# Utilities
numpy
python-dotenv

# other
hf_xet